requirements.txt
pytz
pyyaml
docker
django-generate
jsonschema
redis
websockets
Faker
flake8
autopep8
setuptools
wheel
Deprecated
cryptography
flask
pandas
geopandas
gitpython
python-dotenv
djangorestframework
git+https://github.com/NOAA-OWP/hypy@master#egg=hypy&subdirectory=python
hydrotools.nwis-client
numpy
scikit-learn
minio
uri
aiohttp==3.7.4
channels
channels-redis
Pint
django_rq
requests
Django
scipy
xarray
python-dateutil
h5py
asgiref
pip
attrs
Pillow
Jinja2
click
Fiona
cligj
munch
six
cffi
matplotlib
MarkupSafe
tables
numexpr
pyproj
certifi
Shapely
packaging
pyparsing